some tricks and stuff
trick controls test, as well as some extra little bits. after testing I noticed some side affects with the stabilization script so added an extra solution. after testing I realized stabilization was actually exactly  what it said and so it seems it may be used as an easy mode due to it making tricks far easier to land.